+++ packages/experiment-browser/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.23.0](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-js-client@1.22.1...@amplitude/experiment-js-client@1.23.0) (2026-08-24)
+
+
+### Features
+
+* **experiment-tag:** run web experiment client under pending consent ([#361](https://github.com/amplitude/experiment-js-client/issues/361)) ([8595ab2](https://github.com/amplitude/experiment-js-client/commit/8595ab23b1d8b77b16003ba6da6af9dd78600fde))
+
+
+
+
+
+## [1.22.1](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-js-client@1.22.0...@amplitude/experiment-js-client@1.22.1) (2026-08-19)
+
+
+### Bug Fixes
+
+* upgrade target to es6 ([#367](https://github.com/amplitude/experiment-js-client/issues/367)) ([cd04e86](https://github.com/amplitude/experiment-js-client/commit/cd04e86d4d3678be16c990132164d34a3cd89d53))
+
+
+
+
+
+# [1.22.0](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-js-client@1.21.4...@amplitude/experiment-js-client@1.22.0) (2026-08-11)
+
+
+### Features
+
+* **experiment-browser:** support injectable in-memory cache storage for web experiment ([#355](https://github.com/amplitude/experiment-js-client/issues/355)) ([6acec43](https://github.com/amplitude/experiment-js-client/commit/6acec437e93281cfc3aac15753bfdd071263b89e))
+
+
+
+
+
 ## [1.21.4](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-js-client@1.21.3...@amplitude/experiment-js-client@1.21.4) (2026-07-29)
 
 **Note:** Version bump only for package @amplitude/experiment-js-client